refactor: protect enemy and route from mutability from outside

+++ core/src/main/java/com/jazzjack/rab/bit/Game.java
@@ -32,7 +32,7 @@ public class Game extends ApplicationAdapter {
 
     private void initCamera() {
         camera = new OrthographicCamera();
-        camera.setToOrtho(false, WIDTH * SCALE, HEIGHT * SCALE);
+        camera.setToOrtho(false, (float) WIDTH * SCALE, (float) HEIGHT * SCALE);
         camera.update();
     }
